Patching to the logs
Moved the commands into cogs
Split the verify command from its button, the button was put into the uis folder
Added a .gitignore position for debugging
