+++ .gitignore
@@ -2,3 +2,4 @@
 __pycache__
 .vscode
 verified.json
+settings_temp.py
